Revert "Use collapsible markdown widgets for cells in srcmd (#136)"
This reverts commit 74b1e81900fddbc9098cceba26573304ce655d1b.
