--- packages/api/srcmd/types.mts
@@ -1,14 +1,0 @@
-import { type CellType, type SrcbookMetadataType } from '@srcbook/shared';
-
-export type DecodeErrorResult = {
-  error: true;
-  errors: string[];
-};
-
-export type DecodeSuccessResult = {
-  error: false;
-  cells: CellType[];
-  metadata: SrcbookMetadataType;
-};
-
-export type DecodeResult = DecodeErrorResult | DecodeSuccessResult;
