+++ packages/api/test/srcmd_files/mock_notebook_dir/README.md
@@ -1,13 +1,10 @@
 <!-- srcbook:{"language":"javascript"} -->
 
-# Srcbook
+# Notebook
 
-<details>
-  <summary>package.json</summary>
+###### package.json
 
 [package.json](./package.json)
-
-</details>
 
 With some words right behind it.
 
@@ -17,12 +14,9 @@ With some **bold** text and some _italic_ text.
 
 > And a quote, why the f\*\*\* not!
 
-<details>
-  <summary>foo.mjs</summary>
+###### foo.mjs
 
 [foo.mjs](./src/foo.mjs)
-
-</details>
 
 ```json
 { "simple": "codeblock" }
