+++ packages/api/test/srcmd_files/mock_notebook_dir/src/foo.mjs
